requirements.txt
PyYAML==3.12
cycler==0.10.0
docopt==0.6.2
matplotlib==1.5.3
numpy==1.11.2
pyepics==3.2.6
pyparsing==2.1.10
python-dateutil==2.5.3
pytz==2016.7
pyxdg==0.25
scipy==0.18.1
six==1.10.0
xdg==1.0.3
